+++ pica-core/CHANGELOG.md
@@ -8,5 +8,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Added `ParseResult` type
-* Added `TagRef` type
-* Added `Tag` type
+* Added `TagRef` and `Tag` type
+* Added `OccurrenceRef` and `Occurrence` type